migration to new broken-up views to match new doc format
